+++ small_text/stopping_criteria/__init__.py
@@ -1,10 +1,12 @@
 from small_text.stopping_criteria.base import DeltaFScore
 from small_text.stopping_criteria.change import ClassificationChange
 from small_text.stopping_criteria.kappa import KappaAverage
+from small_text.stopping_criteria.uncertainty import OverallUncertainty
 
 
 __all__ = [
     'ClassificationChange',
     'DeltaFScore',
-    'KappaAverage'
+    'KappaAverage',
+    'OverallUncertainty'
 ]